+++ image_flip/CHANGELOG.rst
@@ -1,3 +1,6 @@
+0.0.2 (2022-05-01)
+------------------
+* Move to rclcpp_components composable node
 0.0.1 (2022-04-03)
 ------------------
 * Original release
